feat: Refactor CartContainer to SalesManagement, update API link, and add new components for item management
