+++ package.json
@@ -28,6 +28,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.54.0",
     "react-icons": "^5.4.0",
+    "react-modal": "^3.16.3",
     "react-spinners": "^0.15.0",
     "stripe": "^16.6.0",
     "zod": "^3.23.8"
@@ -38,6 +39,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/react-modal": "^3.16.3",
     "eslint": "^8",
     "eslint-config-next": "14.2.2",
     "postcss": "^8",
