+++ src/app/globals.css
@@ -5,10 +5,11 @@
 *{
     -webkit-font-smoothing: antialiased;
   }
-  html {
-    scroll-behavior: smooth;
-    overflow-x: hidden;
-  }
+  
+html {
+  scroll-behavior: smooth;
+  overflow-x: hidden;
+}
 
 body{
     background-color:black ;
@@ -17,10 +18,10 @@ body{
 ::selection {
     background-color:black; /* cor de fundo dourada */
     color: white; /* cor do texto durante a seleção, preto neste caso */
-  }
+}
   
-  /* Para compatibilidade com navegadores mais antigos, como o Safari */
-  ::-webkit-selection {
-    background-color: black;
-    color: white;
-  }
+/* Para compatibilidade com navegadores mais antigos, como o Safari */
+::-webkit-selection {
+  background-color: black;
+  color: white;
+}
